[IMP] stock_secondary_unit: Compatibility with product variants TT40323
